+++ src/tree-sitter/ExpressionBuilder.ts
@@ -1,22 +1,22 @@
 import {
-  Expression,
   ExpressionFactory,
   ParameterType,
   ParameterTypeRegistry,
 } from '@cucumber/cucumber-expressions'
-import Parser from 'tree-sitter'
+import { resolve } from 'path'
+import Parser, { SyntaxNode } from 'tree-sitter'
+import { Range } from 'vscode-languageserver-types'
 
 import { getLanguage } from './languages.js'
 import {
   ExpressionBuilderResult,
+  ExpressionLink,
   LanguageName,
   ParameterTypeMeta,
   ParserAdapter,
+  PartialLocationLink,
   Source,
 } from './types.js'
-
-const defineStepDefinitionQueryKeys = <const>['expression']
-const defineParameterTypeKeys = <const>['name', 'expression']
 
 export class ExpressionBuilder {
   constructor(private readonly parserAdapter: ParserAdapter) {}
@@ -25,7 +25,7 @@ export class ExpressionBuilder {
     sources: readonly Source<LanguageName>[],
     parameterTypes: readonly ParameterTypeMeta[]
   ): ExpressionBuilderResult {
-    const expressions: Expression[] = []
+    const expressionLinks: ExpressionLink[] = []
     const errors: Error[] = []
     const registry = new ParameterTypeRegistry()
     const expressionFactory = new ExpressionFactory(registry)
@@ -66,13 +66,15 @@ export class ExpressionBuilder {
       for (const defineParameterTypeQuery of treeSitterLanguage.defineParameterTypeQueries) {
         const query = this.parserAdapter.query(defineParameterTypeQuery)
         const matches = query.matches(tree.rootNode)
-        const records = matches.map((match) => recordFromMatch(match, defineParameterTypeKeys))
-        for (const record of records) {
-          const name = record['name']
-          const regexp = record['expression']
-          if (name && regexp) {
+        for (const match of matches) {
+          const nameNode = syntaxNode(match, 'name')
+          const regexpNode = syntaxNode(match, 'expression')
+          if (nameNode && regexpNode) {
             defineParameterType(
-              makeParameterType(toString(name), treeSitterLanguage.toStringOrRegExp(regexp))
+              makeParameterType(
+                toString(nameNode.text),
+                treeSitterLanguage.toStringOrRegExp(regexpNode.text)
+              )
             )
           }
         }
@@ -90,15 +92,24 @@ export class ExpressionBuilder {
       for (const defineStepDefinitionQuery of treeSitterLanguage.defineStepDefinitionQueries) {
         const query = this.parserAdapter.query(defineStepDefinitionQuery)
         const matches = query.matches(tree.rootNode)
-        const records = matches.map((match) =>
-          recordFromMatch(match, defineStepDefinitionQueryKeys)
-        )
-        for (const record of records) {
-          const expression = record['expression']
-          if (expression) {
-            const stringOrRegexp = treeSitterLanguage.toStringOrRegExp(expression)
+        for (const match of matches) {
+          const expressionNode = syntaxNode(match, 'expression')
+          if (expressionNode) {
+            const stringOrRegexp = treeSitterLanguage.toStringOrRegExp(expressionNode.text)
             try {
-              expressions.push(expressionFactory.createExpression(stringOrRegexp))
+              const expression = expressionFactory.createExpression(stringOrRegexp)
+              const targetRange: Range = Range.create(
+                expressionNode.startPosition.row,
+                expressionNode.startPosition.column,
+                expressionNode.endPosition.row,
+                expressionNode.endPosition.column
+              )
+              const partialLink: PartialLocationLink = {
+                targetRange,
+                targetSelectionRange: targetRange,
+                targetUri: `file://${resolve(source.path)}`,
+              }
+              expressionLinks.push({ expression, partialLink })
             } catch (err) {
               errors.push(err)
             }
@@ -108,7 +119,7 @@ export class ExpressionBuilder {
     }
 
     return {
-      expressions,
+      expressionLinks,
       errors,
       registry,
     }
@@ -121,15 +132,8 @@ function toString(s: string): string {
   return match[1]
 }
 
-function recordFromMatch<T extends string>(
-  match: Parser.QueryMatch,
-  keys: readonly T[]
-): Record<T, string | undefined> {
-  const values = keys.map((name) => match.captures.find((c) => c.name === name)?.node?.text)
-  return Object.fromEntries(keys.map((_, i) => [keys[i], values[i]])) as Record<
-    T,
-    string | undefined
-  >
+function syntaxNode(match: Parser.QueryMatch, name: string): SyntaxNode | undefined {
+  return match.captures.find((c) => c.name === name)?.node
 }
 
 function makeParameterType(name: string, regexp: string | RegExp) {