+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.21.0] - 2022-05-23
+### Added
+- Implement go to step definition ([#48](https://github.com/cucumber/language-service/pull/48))
+
 ## [0.20.4] - 2022-05-12
 ### Fixed
 - Fix release.
@@ -172,7 +176,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ([#1732](https://github.com/cucumber/common/pull/1732)
 [aslakhellesoy](https://github.com/aslakhellesoy))
 
-[Unreleased]: https://github.com/cucumber/language-service/compare/v0.20.4...HEAD
+[Unreleased]: https://github.com/cucumber/language-service/compare/v0.21.0...HEAD
+[0.21.0]: https://github.com/cucumber/language-service/compare/v0.20.4...v0.21.0
 [0.20.4]: https://github.com/cucumber/language-service/compare/v0.20.3...v0.20.4
 [0.20.3]: https://github.com/cucumber/language-service/compare/v0.20.2...v0.20.3
 [0.20.2]: https://github.com/cucumber/language-service/compare/v0.20.1...v0.20.2